refactor(usage-insights): extract shared transcript helpers into scripts/lib
All three collectors independently reimplemented transcript discovery
(~/.claude/projects walk, including nested subagents/**), tool-error
bucketing, and counting-map helpers. Pull the common pieces into
scripts/lib/transcripts.mjs and import them.

- new scripts/lib/transcripts.mjs: PROJECTS_DIR, encodeCwd,
  resolveProjectDir, transcriptFilesIn, allTranscriptFiles,
  categorizeToolError, bump, mergeCounts, mapToSortedList
- collect-usage.mjs / collect-insights-data.mjs: drop local duplicates,
  import from the lib; declare attributedRecords (was an undeclared
  assignment, a live ReferenceError under ESM strict mode)
- collect-onboarding-data.mjs: share PROJECTS_DIR/resolveProjectDir but
  keep its own scanner, which reads whole files and walks only the top
  level; comment explains why
- trim facets.md / suggestions.md schemas to the fields the renderer
  actually consumes, and align skills/insights/SKILL.md prose with them
- README: document the new lib in the scripts table and file tree
